Move deleter into fsutil
